Add an RNG module
This can encapsulate random number generation for our purposes. We just
have a single interface describing what a random number generation is,
as well as a default implementation of that interface. Doing it this way
makes it easy to change the implementation of random number generation
later very easily.

Right now I went with a simple linear PRNG, but we could do something
fancy like a mersenne twister or something later.